first attempt to reorganize directory
