+++ nwb_conversion_tools/datainterfaces/ophys/imagingextractorinterface.py
@@ -1,6 +1,6 @@
 from roiextractors import SbxImagingExtractor, Hdf5ImagingExtractor, \
     TiffImagingExtractor
-from ..baseimagingextractorinterface import BaseImagingExtractorInterface
+from ...baseimagingextractorinterface import BaseImagingExtractorInterface
 
 
 class TiffImagingInterface(BaseImagingExtractorInterface):
