+++ nwb_conversion_tools/datainterfaces/ophys/roiextractordatainterface.py
@@ -1,7 +1,7 @@
 from roiextractors import CnmfeSegmentationExtractor, ExtractSegmentationExtractor, \
     CaimanSegmentationExtractor, Suite2pSegmentationExtractor, SimaSegmentationExtractor
 
-from ..basesegmentationextractorinterface import BaseSegmentationExtractorInterface
+from ...basesegmentationextractorinterface import BaseSegmentationExtractorInterface
 
 
 class CnmfeSegmentationInterface(BaseSegmentationExtractorInterface):
